Integrate sqlboiler

+++ domain/user/userRepository.go
@@ -1,6 +1,6 @@
 package usr
 
 type UserRepository interface {
-	CreateUser(user User) (int, error)
+	CreateUser(user User) (int64, error)
 	GetUserByEmail(email string) (*User, error)
 }
